Add forest background

+++ wtf/poly.py
@@ -17,25 +17,32 @@ class RockPoly:
         gl.GL_ONE_MINUS_SRC_ALPHA,
     )
 
-    def __init__(self, verts):
+    def __init__(self, verts, draw=True):
         self.indexes = earcut(verts)
-        self.dl = self.batch.add_indexed(
-            len(verts) // 2,
-            gl.GL_TRIANGLES,
-            self.group,
-            self.indexes,
-            ('v2f/static', np.array(verts) / SPACE_SCALE),
-            ('t2f/static', np.array(verts) * 0.1)
-        )
+
+        if draw:
+            self.dl = self.batch.add_indexed(
+                len(verts) // 2,
+                gl.GL_TRIANGLES,
+                self.group,
+                self.indexes,
+                ('v2f/static', np.array(verts) / SPACE_SCALE),
+                ('t2f/static', np.array(verts) / (512 * SPACE_SCALE * 2))
+            )
+        else:
+            self.dl = None
 
         self.shapes = []
         verts = np.array(verts)
         tris = verts.reshape(-1, 2)[self.indexes].reshape(-1, 3, 2)
         for tri in tris:
             shp = Poly(space.static_body, tri)
+            shp.friction = 0.6
+            shp.elasticity = 0.6
             space.add(shp)
             self.shapes.append(shp)
 
     def delete(self):
-        self.dl.delete()
+        if self.dl:
+            self.dl.delete()
         space.remove(*self.shapes)
